Pull up `TextOutput` and `AudioOutput` into abstract classes users can inherit.  (#445)
